requirements.txt
Flask
gunicorn
pandas
python-dotenv
